fix misc issues
tableWidgetExtension.ts now types injected services as MarkdownRenderService and LinkOpener, not ReturnType`<typeof ...>`. Moved LinkOpener and createLinkOpener into linkOpener.ts. joplinBridge.ts is now just the postMessage bridge. Added the clear() comment: in-flight requests still resolve for existing callbacks, but their results are not cached.
